SWITCHYARD-1220: Follow-up work post-Drools/jBPM 6 upgrade
