+++ demos/helpdesk-webapp/src/main/java/org/switchyard/quickstarts/demos/helpdesk/HelpDeskServiceProcess.java
@@ -18,21 +18,24 @@
  */
 package org.switchyard.quickstarts.demos.helpdesk;
 
+//import org.jbpm.process.audit.JPAWorkingMemoryDbLogger;
 import org.jbpm.process.workitem.wsht.MinaHTWorkItemHandler;
 import org.switchyard.component.bpm.annotation.BPM;
 import org.switchyard.component.bpm.annotation.StartProcess;
 import org.switchyard.component.bpm.annotation.WorkItemHandler;
 import org.switchyard.component.bpm.work.SwitchYardServiceWorkItemHandler;
+import org.switchyard.component.common.knowledge.KnowledgeConstants;
+//import org.switchyard.component.common.knowledge.annotation.Listener;
 import org.switchyard.component.common.knowledge.annotation.Manifest;
 import org.switchyard.component.common.knowledge.annotation.Mapping;
 import org.switchyard.component.common.knowledge.annotation.Resource;
-import org.switchyard.component.common.knowledge.util.Mappings;
 
 /**
  * @author David Ward &lt;<a href="mailto:dward@jboss.org">dward@jboss.org</a>&gt; &copy; 2012 Red Hat Inc.
  */
 @BPM(value=HelpDeskService.class,
     persistent=true,
+    //listeners={@Listener(JPAWorkingMemoryDbLogger.class)},
     manifest=@Manifest(resources=@Resource(location="/META-INF/HelpDeskService.bpmn", type="BPMN2")),
     workItemHandlers={@WorkItemHandler(SwitchYardServiceWorkItemHandler.class), @WorkItemHandler(MinaHTWorkItemHandler.class)})
 public interface HelpDeskServiceProcess extends HelpDeskService {
@@ -40,7 +43,7 @@ public interface HelpDeskServiceProcess extends HelpDeskService {
     @Override
     @StartProcess(
         inputs={@Mapping(expression="message.content", variable="ticket")},
-        outputs={@Mapping(expression="ticketAck", variable=Mappings.CONTENT_OUTPUT)}
+        outputs={@Mapping(expression="ticketAck", variable=KnowledgeConstants.CONTENT_OUTPUT)}
     )
     public TicketAck openTicket(Ticket ticket);
