+++ pom.xml
@@ -60,7 +60,7 @@
         <module>rest-binding</module>
         <module>rules-camel-cbr</module>
         <module>rules-interview</module>
-        <module>rules-interview-agent</module>
+        <module>rules-interview-container</module>
         <module>jca-inflow-hornetq</module>
         <module>jca-outbound-hornetq</module>
         <module>transform-jaxb</module>
