--- rules-interview-agent/src/main/resources/org/switchyard/quickstarts/rules/interview/Interview-21.drl
@@ -1,37 +1,0 @@
-# JBoss, Home of Professional Open Source
-# Copyright 2012 Red Hat Inc. and/or its affiliates and other contributors
-# as indicated by the @authors tag. All rights reserved.
-# See the copyright.txt in the distribution for a
-# full listing of individual contributors.
-#
-# This copyrighted material is made available to anyone wishing to use,
-# modify, copy, or redistribute it subject to the terms and conditions
-# of the GNU Lesser General Public License, v. 2.1.
-# This program is distributed in the hope that it will be useful, but WITHOUT A
-# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
-# PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more details.
-# You should have received a copy of the GNU Lesser General Public License,
-# v.2.1 along with this distribution; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
-# MA  02110-1301, USA.
-
-package org.switchyard.quickstarts.rules.interview
-
-global java.lang.String service
-
-rule "Is of valid age"
-    when
-        $a : Applicant( age > 20 )
-    then
-        $a.setValid( true );
-        System.out.println("********** " + service + ": " + $a.getName() + " is a valid applicant. **********");
-end
-
-rule "Is not of valid age"
-    when
-        $a : Applicant( age < 21 )
-    then
-        $a.setValid( false );
-        System.err.println("********** " + service + ": " + $a.getName() + " is NOT a valid applicant. **********");
-end
-
