+++ rules-interview-container/src/main/java/org/switchyard/quickstarts/rules/interview/Applicant.java
