+++ rules-interview-container/src/main/java/org/switchyard/quickstarts/rules/interview/InterviewRules.java
@@ -18,9 +18,9 @@
  */
 package org.switchyard.quickstarts.rules.interview;
 
+import org.switchyard.component.common.knowledge.annotation.Container;
 import org.switchyard.component.common.knowledge.annotation.Manifest;
 import org.switchyard.component.common.knowledge.annotation.Mapping;
-import org.switchyard.component.common.knowledge.annotation.Resource;
 import org.switchyard.component.rules.annotation.Execute;
 import org.switchyard.component.rules.annotation.Rules;
 
@@ -28,10 +28,8 @@ import org.switchyard.component.rules.annotation.Rules;
  * @author David Ward &lt;<a href="mailto:dward@jboss.org">dward@jboss.org</a>&gt; &copy; 2012 Red Hat Inc.
  */
 @Rules(value=Interview.class,
-    manifest=@Manifest(resources={@Resource(location=InterviewRules.PKG, type="PKG")}, scan=false)) // TODO: scan=true
+    manifest=@Manifest(container=@Container(sessionName="rules-interview-container")))
 public interface InterviewRules extends Interview {
-
-    static String PKG = "org/switchyard/quickstarts/rules/interview/Interview.pkg";
 
     @Override
     @Execute(globals=@Mapping(expression="exchange.provider.name.localPart", variable="service"))
