+++ rules-interview-container/src/main/resources/META-INF/kmodule.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<kmodule xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://jboss.org/kie/6.0.0/kmodule">
+    <kbase name="org.switchyard.quickstarts.rules.interview">
+        <ksession name="rules-interview-container"/>
+    </kbase>
+</kmodule>
