+++ rules-interview-container/src/main/resources/META-INF/switchyard.xml
@@ -3,8 +3,8 @@
             xmlns:swyd="urn:switchyard-config:switchyard:1.0"
             xmlns:rules="urn:switchyard-component-rules:config:1.0"
             xmlns:sca="http://docs.oasis-open.org/ns/opencsa/sca/200912"
-            name="rules-interview-agent">
-    <sca:composite name="Interview" targetNamespace="urn:switchyard-quickstart:rules-interview-agent:0.1.0">
+            name="rules-interview-container">
+    <sca:composite name="Interview" targetNamespace="urn:switchyard-quickstart:rules-interview-container:0.1.0">
         <sca:service name="Interview" promote="Interview">
             <sca:interface.java interface="org.switchyard.quickstarts.rules.interview.Interview"/>
         </sca:service>
