+++ rules-interview-container/src/main/resources/org/switchyard/quickstarts/rules/interview/Interview.drl
